test(extra): centralize agent integration tests
Centralize agent coverage in the extra suite and build the binary once.
