+++ crates/ironrdp-testsuite-extra/tests/main.rs
@@ -1,4 +1,6 @@
 #![allow(unused_crate_dependencies)] // false positives because there is both a library and a binary
+#![allow(clippy::panic, reason = "panic is fine in tests")]
+#![allow(clippy::std_instead_of_core, reason = "std is fine in integration tests")]
 #![allow(clippy::unwrap_used, reason = "unwrap is fine in tests")]
 
 #[cfg(feature = "full")]
